chore: put all Combo box components into separate files
