+++ src/components/ComboBox/ComboBoxCollection.tsx
@@ -1,0 +1,7 @@
+import { Combobox as ComboboxPrimitive } from '@base-ui/react';
+
+const ComboboxCollection = ({ ...props }: ComboboxPrimitive.Collection.Props) => {
+  return <ComboboxPrimitive.Collection data-slot="combobox-collection" {...props} />;
+};
+
+export { ComboboxCollection };
